lite/test_int8_model/requirements.txt
patchelf
numpy
opencv-python
paddlepaddle>=2.4.0
paddledet>=2.5.0
paddleseg==2.5.0
paddlenlp>=2.3.0
pandas==0.23.4
